Fix CUDA training: convert tensors to int64 (long) for cross_entropy
PROBLEM: NotImplementedError: nll_loss_forward_reduce_cuda_kernel_2d_index not implemented for 'Int'
ROOT CAUSE: Data is int32 but PyTorch CUDA cross_entropy requires int64

SOLUTION: Add .long() conversion in get_batch()
- x and y tensors now int64 before moving to CUDA
- Compatible with PyTorch's cross_entropy loss function

+++ train.py
@@ -338,8 +338,9 @@ def get_batch(data, batch_size, block_size, device):
     # Stack as numpy array first for faster conversion
     x = np.stack([data[i:i+block_size] for i in ix])
     y = np.stack([data[i+1:i+block_size+1] for i in ix])
-    x = torch.from_numpy(x).to(device)
-    y = torch.from_numpy(y).to(device)
+    # Convert to long (int64) for PyTorch CUDA compatibility
+    x = torch.from_numpy(x).long().to(device)
+    y = torch.from_numpy(y).long().to(device)
     return x, y
 
 @torch.no_grad()
